requirements.txt
asgiref==3.7.2
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.1
djangorestframework==3.14.0
filelock==3.14.0
identify==2.5.36
idna==3.7
nodeenv==1.9.0
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.12.1
urllib3==2.2.1
virtualenv==20.26.2
